Conduct ensemble of "corrected" basins

+++ build_qf_sim_from_json.py
@@ -4,7 +4,8 @@ from shutil import copy
 
 qf_dir = Path("D:/Fire-Flood-Mud/QF_runs/SBS")
 
-fires = ["Caldor", "CedarCreek", "CubCreek2", "Dixie", "KNP"]
+# fires = ["Caldor", "CedarCreek", "CubCreek2", "Dixie", "KNP"]
+fires = ["CedarCreek", "CubCreek2"]
 qf_path = (
     Path(__file__).parent.parent.parent
     / "Quicfire"
@@ -14,14 +15,16 @@ qf_path = (
 )
 
 for fire in fires:
-    sites = [f"{fire[:3]}{i}" for i in range(1, 21)]
+    # sites = [f"{fire[:3]}{i}" for i in range(1, 21)]
+    sites = [f"{fire[:3]}{i}_COR" for i in range(1, 4)]
     for site in sites:
-        print(f"{fire} - {site}")
         run_path = qf_dir / fire / site
-        json_path = run_path / f"{site}.json"
-        sim = SimulationInputs.from_json(json_path)
-        sim.write_inputs(run_path)
-        dst = run_path / "quicfire_LIN64.exe"
-        copy(qf_path, dst)
-        mac_path = run_path / "quicfire_MACI.exe"
-        mac_path.unlink(missing_ok=True)
+        if run_path.exists():
+            print(f"{fire} - {site}")
+            json_path = run_path / f"{site}.json"
+            sim = SimulationInputs.from_json(json_path)
+            sim.write_inputs(run_path)
+            dst = run_path / "quicfire_LIN64.exe"
+            copy(qf_path, dst)
+            mac_path = run_path / "quicfire_MACI.exe"
+            mac_path.unlink(missing_ok=True)
